gemini/webgem/python/requirements.txt
flask
mistune
pillow
google-genai
